Modified contract to work with native token 🪙 (#32)
Modified the smart contract to not use ERC-20 and instead use the native
blockchain token.

This removes some complication of methods and allows to a more
streamline implementation.

Updated CI to deploy with this new parameters.

This pull request simplifies the `Raffle` contract by removing the
dependency on `ERC20` tokens and switching to using `ETH` for ticket
purchases. It also updates the related deployment scripts and tests to
reflect these changes.

## Summary

- **New Features**
- Transitioned ticket purchases from ERC20 tokens to direct Ether
payments.
- Updated ticket price input to specify Ether units in the deployment
process.

- **Bug Fixes**
- Simplified the ticket purchasing process by removing unnecessary
token-related logic.

- **Documentation**
	- Clarified README instructions regarding ticket pricing in ETH.
- Removed outdated information about the Raffle Token from the
documentation.

- **Chores**
- Deleted the RaffleToken contract and associated deployment module,
streamlining the codebase.